requirements.txt
Flask==3.0.3
transformers==4.45.1
faiss-cpu==1.8.0.post1
torch==2.2.2
numpy==1.26.4
gunicorn==23.0.0
